pyproject.toml
[tool.poetry]
name = "wisdm-activity-recognition"
version = "0.1.0"
description = "Activity recognition using WISDM dataset"
authors = ["Andrew Shuleshko <shuleshko.av@phystech.edu>"]
packages = [{include = "wisdm"}]

[tool.poetry.dependencies]
python = "^3.10"
torch = "^2.2.0"
pytorch-lightning = "^2.2.0"
hydra-core = "^1.3.2"
mlflow = "^2.10.0"
pandas = "^2.2.0"
numpy = "^1.26.0"
scikit-learn = "^1.4.0"
onnx = "^1.15.0"
onnxruntime = "^1.17.0"
rich = "^13.7.0"
seaborn = "^0.13.2"

[tool.poetry.group.dev.dependencies]
black = "^24.2.0"
isort = "^5.13.0"
flake8 = "^7.0.0"
pytest = "^8.0.0"
pytest-cov = "^4.1.0"
pre-commit = "^3.6.0"
jupyterlab = "^4.1.0"
wandb = "^0.16.0"

[tool.black]
line-length = 88
target-version = ["py310"]

[tool.isort]
profile = "black"
multi_line_output = 3

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
